adds ModelWorkspaceCTAs.svelte
